backend/api/lexical/requirements.txt
Django
django-cors-headers
django-filter
djangorestframework
gunicorn
whitenoise